pyproject.toml
[tool.poetry]
name = "wowadv"
version = "0.1.0"
description = ""
authors = ["Olga <data19101988@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2 = "^2.9.6"
django-bootstrap-datepicker-plus = "^5.0.4"
install = "^1.3.5"
django-bootstrap5 = "^23.3"
l = "^0.11.0"
django-ckeditor-5 = "^0.2.8"
django-allauth = "^0.54.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
